해커톤 개발 최종
1. 카카오페이 리다이렉트로 수정
2. 활동 내역 조회 추가
3. 이미지 한 게시글에 하나만 가능
4. 질문에 이미지 추가

+++ likelion/src/main/java/com/power/likelion/service/AnswerSerivce.java
@@ -93,6 +93,11 @@ public class AnswerSerivce {
             throw new Exception("답변 작성자가 아닙니다.");
         }
 
+        if(answer.getAnswerCheck().equals(CheckStatus.True)){
+           throw new Exception("답변이 채택되어 삭제할 수 없습니다.");
+        }
+
+
 
         answerRepository.delete(answer);
     }
